refactor evaluate/plotting (#2287)
* Refactor: extract plot classes from plotter.py into separate modules

- Remove BarPlots, LinePlots, QuantilePlots, ScoreCards classes from plotter.py (already in separate files from #2140)
- Move utility functions (format_datetime, calculate_average_over_dim, etc.) to plot_utils.py
- Update plot_orchestration.py to import from separate modules
- Add plotting __init__.py
- plotter.py now contains only the Plotter base class

* add new files
