+++ packages/evaluate/src/weathergen/evaluate/plotting/plotter.py
@@ -1,7 +1,6 @@
 import datetime
 import logging
 import os
-import re
 import warnings
 from pathlib import Path
 
@@ -11,16 +10,13 @@ import matplotlib as mpl
 import matplotlib.pyplot as plt
 import numpy as np
 import omegaconf as oc
-import seaborn as sns
 import xarray as xr
 from astropy_healpix import HEALPix as HEALPixGrid
 from cartopy.io import DownloadWarning
 from matplotlib.collections import LineCollection
-from matplotlib.lines import Line2D
-from scipy.stats import wilcoxon
 
 from weathergen.common.config import _load_private_conf
-from weathergen.evaluate.plotting.plot_utils import DefaultMarkerSize
+from weathergen.evaluate.plotting.plot_utils import DefaultMarkerSize, format_datetime
 from weathergen.evaluate.utils.regions import RegionBoundingBox
 
 _logger = logging.getLogger(__name__)
@@ -632,1361 +628,3 @@ class Plotter:
                 title += f" ({format_datetime(valid_time_start)})"
 
         return title
-
-
-class LinePlots:
-    def __init__(self, plotter_cfg: dict, output_basedir: str | Path):
-        """
-        Initialize the LinePlots class.
-
-        Parameters
-        ----------
-        plotter_cfg:
-            Configuration dictionary containing basic information for plotting.
-            Expected keys are:
-                - image_format: Format of the saved images (e.g., 'png', 'pdf', etc.)
-                - dpi_val: DPI value for the saved images
-                - fig_size: Size of the figure (width, height) in inches
-                -  plot_ensemble:
-                    If True, plot ensemble spread if 'ens' dimension is present. Options are:
-                        - False: do not plot ensemble spread
-                        - "std": plot mean +/- standard deviation
-                        - "minmax": plot min-max range
-                        - "members": plot individual ensemble members
-        output_basedir:
-            Base directory under which the plots will be saved.
-            Expected scheme `<results_base_dir>/<run_id>`.
-        """
-
-        self.image_format = plotter_cfg.get("image_format")
-        self.dpi_val = plotter_cfg.get("dpi_val")
-        self.fig_size = plotter_cfg.get("fig_size")
-        self.log_scale = plotter_cfg.get("log_scale")
-        self.add_grid = plotter_cfg.get("add_grid")
-        self.plot_ensemble = plotter_cfg.get("plot_ensemble", False)
-        self.baseline = plotter_cfg.get("baseline")
-        self.out_plot_dir = Path(output_basedir) / "line_plots"
-        if not os.path.exists(self.out_plot_dir):
-            _logger.info(f"Creating dir {self.out_plot_dir}")
-            os.makedirs(self.out_plot_dir, exist_ok=True)
-
-        _logger.info(f"Saving summary plots to: {self.out_plot_dir}")
-
-    def _check_lengths(self, data: xr.DataArray | list, labels: str | list) -> tuple[list, list]:
-        """
-        Check if the lengths of data and labels match.
-
-        Parameters
-        ----------
-        data:
-            DataArray or list of DataArrays to be plotted
-        labels:
-            Label or list of labels for each dataset
-
-        Returns
-        -------
-            data_list, label_list - lists of data and labels
-        """
-        assert isinstance(data, xr.DataArray | list), (
-            "Compare::plot - Data should be of type xr.DataArray or list"
-        )
-        assert isinstance(labels, str | list), (
-            "Compare::plot - Labels should be of type str or list"
-        )
-
-        # convert to lists
-
-        data_list = [data] if isinstance(data, xr.DataArray) else data
-        label_list = [labels] if isinstance(labels, str) else labels
-
-        assert len(data_list) == len(label_list), "Compare::plot - Data and Labels do not match"
-
-        return data_list, label_list
-
-    def print_all_points_from_graph(self, fig: plt.Figure) -> None:
-        for ax in fig.get_axes():
-            for line in ax.get_lines():
-                ydata = line.get_ydata()
-                xdata = line.get_xdata()
-                label = line.get_label()
-                _logger.info(f"Summary for {label} plot:")
-                for xi, yi in zip(xdata, ydata, strict=False):
-                    xi = xi if isinstance(xi, str) else f"{float(xi):.3f}"
-                    yi = yi if isinstance(yi, str) else f"{float(yi):.3f}"
-                    _logger.info(f"  x: {xi}, y: {yi}")
-                _logger.info("--------------------------")
-        return
-
-    def _plot_ensemble(self, data: xr.DataArray, x_dim: str, label: str) -> None:
-        """
-        Plot ensemble spread for a data array.
-
-        Parameters
-        ----------
-        data: xr.xArray
-            DataArray to be plotted
-        x_dim: str
-            Dimension to be used for the x-axis.
-        label: str
-            Label for the dataset
-        Returns
-        -------
-            None
-        """
-        averaged = data.mean(dim=[dim for dim in data.dims if dim != x_dim], skipna=True).sortby(
-            x_dim
-        )
-
-        lines = plt.plot(
-            averaged[x_dim],
-            averaged.values,
-            label=label,
-            marker="o",
-            linestyle="-",
-        )
-        line = lines[0]
-        color = line.get_color()
-
-        ens = data.mean(
-            dim=[dim for dim in data.dims if dim not in [x_dim, "ens"]], skipna=True
-        ).sortby(x_dim)
-
-        if self.plot_ensemble == "std":
-            std_dev = ens.std(dim="ens", skipna=True).sortby(x_dim)
-            plt.fill_between(
-                averaged[x_dim],
-                (averaged - std_dev).values,
-                (averaged + std_dev).values,
-                label=f"{label} - std dev",
-                color=color,
-                alpha=0.2,
-            )
-
-        elif self.plot_ensemble == "minmax":
-            ens_min = ens.min(dim="ens", skipna=True).sortby(x_dim)
-            ens_max = ens.max(dim="ens", skipna=True).sortby(x_dim)
-
-            plt.fill_between(
-                averaged[x_dim],
-                ens_min.values,
-                ens_max.values,
-                label=f"{label} - min max",
-                color=color,
-                alpha=0.2,
-            )
-
-        elif self.plot_ensemble == "members":
-            for j in range(ens.ens.size):
-                plt.plot(
-                    ens[x_dim],
-                    ens.isel(ens=j).values,
-                    color=color,
-                    alpha=0.2,
-                )
-        else:
-            _logger.warning(
-                f"LinePlot:: Unknown option for plot_ensemble: {self.plot_ensemble}. "
-                "Skipping ensemble plotting."
-            )
-
-    def _preprocess_data(
-        self, data: xr.DataArray, x_dim: str | list[str], verbose: bool = True
-    ) -> xr.DataArray:
-        """
-        Average all dimensions except x_dim (which may be a string or list)
-        and then sort the result.
-
-        Parameters
-        ----------
-        data : xr.DataArray
-            DataArray to be preprocessed.
-        x_dim : str or list of str
-            Dimension(s) to be preserved for the x-axis.
-        verbose : bool
-            Log information about averaging.
-
-        Returns
-        -------
-        xr.DataArray
-            Preprocessed DataArray.
-        """
-
-        x_dims = [x_dim] if isinstance(x_dim, str) else list(x_dim)
-
-        non_x_dims = [dim for dim in data.dims if dim not in x_dims]
-
-        if any(data.sizes.get(dim, 1) > 1 for dim in non_x_dims) and verbose:
-            logging.info(f"Averaging over dimensions: {non_x_dims}")
-
-        out = data.mean(dim=non_x_dims, skipna=True)
-
-        for xd in x_dims:
-            out = out.sortby(xd)
-
-        return out
-
-    def plot(
-        self,
-        data: xr.DataArray | list,
-        labels: str | list,
-        tag: str = "",
-        x_dim: str = "lead_time",
-        y_dim: str = "value",
-        print_summary: bool = False,
-    ) -> None:
-        """
-        Plot a line graph comparing multiple datasets.
-
-        Parameters
-        ----------
-        data:
-            DataArray or list of DataArrays to be plotted
-        labels:
-            Label or list of labels for each dataset
-        tag:
-            Tag to be added to the plot title and filename
-        x_dim:
-            Dimension to be used for the x-axis. The code will average over all other dimensions.
-        y_dim:
-            Name of the dimension to be used for the y-axis.
-        print_summary:
-            If True, print a summary of the values from the graph.
-        Returns
-        -------
-            None
-        """
-
-        data_list, label_list = self._check_lengths(data, labels)
-
-        assert x_dim in data_list[0].dims or x_dim in data_list[0].coords, (
-            f"x dimension '{x_dim}' not found in data dimensions "
-            f"{data_list[0].dims} or coords {data_list[0].coords}."
-        )
-
-        fig = plt.figure(figsize=(12, 6), dpi=self.dpi_val)
-
-        for i, data in enumerate(data_list):
-            non_zero_dims = [dim for dim in data.dims if dim != x_dim and data[dim].shape[0] > 1]
-
-            if self.plot_ensemble and "ens" in non_zero_dims:
-                _logger.info(f"LinePlot:: Plotting ensemble with option {self.plot_ensemble}.")
-                self._plot_ensemble(data, x_dim, label_list[i])
-            else:
-                averaged = self._preprocess_data(data, x_dim)
-
-                plt.plot(
-                    averaged[x_dim],
-                    averaged.values,
-                    label=label_list[i],
-                    marker="o",
-                    linestyle="-",
-                )
-
-        parts = ["compare", tag]
-        name = "_".join(filter(None, parts))
-
-        # TODO: generalise this for other x_dims by instroducing a "units"
-        # entry in the function if needed
-        xunits = "hr" if x_dim == "lead_time" else None
-        self._plot_base(fig, name, x_dim, y_dim, print_summary, xunits=xunits)
-
-    def _plot_base(
-        self,
-        fig: plt.Figure,
-        name: str,
-        x_dim: str,
-        y_dim: str,
-        print_summary: bool = False,
-        line: float | None = None,
-        vlines: bool = False,
-        title: str | None = None,
-        xunits: str | None = None,
-        yunits: str | None = None,
-    ) -> None:
-        """
-        Apply labels, title, legend, save and optionally print summary.
-        Parameters
-        ----------
-        fig:
-            Matplotlib figure to be finalized
-        name:
-            Name of the plot file
-        x_dim:
-            Label for the x-axis
-        y_dim:
-            Label for the y-axis
-        print_summary:
-            If True, print a summary of the values from the graph.
-        line:
-            If provided, draw a horizontal line at the given y-value.
-        vlines:
-            If True, draw vertical lines to separate each group of variables.
-        title:
-            Title for the plot.
-        xunits:
-            Units for the x-axis.
-        yunits:
-            Units for the y-axis.
-        Returns
-        -------
-            None
-        """
-
-        plt.xlabel(
-            "".join(c if c.isalnum() else " " for c in x_dim) + (f" [{xunits}]" if xunits else "")
-        )
-        plt.ylabel(
-            "".join(c if c.isalnum() else " " for c in y_dim) + (f" [{yunits}]" if yunits else "")
-        )
-
-        plt.title(title if title is not None else " ".join(c if c.isalnum() else " " for c in name))
-        plt.legend(frameon=False)
-
-        if self.add_grid:
-            plt.grid(True, linestyle="--", color="gray", alpha=0.5)
-
-        if self.log_scale:
-            plt.yscale("log")
-
-        if print_summary:
-            _logger.info(f"Summary values for {name}")
-            self.print_all_points_from_graph(fig)
-
-        if line:
-            plt.axhline(y=line, color="black", linestyle="--", linewidth=1, zorder=1)
-
-        if vlines:
-            vlines = []
-            last_prefix = None
-
-            channels = [t.get_text() for t in fig.gca().get_xticklabels() if t.get_text()]
-
-            for idx, ch in enumerate(channels):
-                m = re.match(r"([a-zA-Z]+)_\d+", ch)
-                prefix = m.group(1) if m else ch
-                if last_prefix is not None and prefix != last_prefix:
-                    vlines.append(idx - 0.5)
-                last_prefix = prefix
-            for vl in vlines:
-                plt.axvline(x=vl, color="#001f3f", linestyle="-", linewidth=0.5, zorder=1)
-
-        plt.tight_layout()
-        plt.savefig(f"{self.out_plot_dir.joinpath(name)}.{self.image_format}")
-        plt.close()
-
-    def ratio_plot(
-        self,
-        data: xr.DataArray | list,
-        run_ids: list[str],
-        labels: str | list,
-        tag: str = "",
-        x_dim: str = "forecast_step",
-        y_dim: str = "value",
-        print_summary: bool = False,
-    ) -> None:
-        """
-        Plot a ratio plot comparing multiple datasets to the first dataset.
-        Parameters
-        ----------
-        data:
-            DataArray or list of DataArrays to be plotted
-        run_ids:
-            List of run IDs corresponding to each dataset
-        labels:
-            Label or list of labels for each dataset
-        tag:
-            Tag to be added to the plot title and filename
-        x_dim:
-            Dimension to be used for the x-axis. The code will average over all other dimensions.
-        y_dim:
-            Name of the dimension to be used for the y-axis.
-        print_summary:
-            If True, print a summary of the values from the graph.
-        Returns
-        -------
-            None
-        """
-
-        data_list, label_list = self._check_lengths(data, labels)
-
-        if len(data_list) < 2:
-            baseline = xr.full_like(data_list[0], 1.0)
-            baseline_name = "ones"
-            descr = "scores"
-        else:
-            descr = "ratio_plot"
-            baseline_name = self.baseline
-            baseline_idx = run_ids.index(self.baseline) if self.baseline in run_ids else None
-            if baseline_idx is not None:
-                _logger.info(f"Using baseline run ID '{self.baseline}' for ratio plot.")
-                baseline = data_list[baseline_idx]
-
-            else:
-                baseline_name = run_ids[0]
-                baseline = data_list[0]
-
-        ref_raw = self._preprocess_data(baseline, x_dim, verbose=False)
-
-        channel_names = set(ref_raw.channel.values)
-        # Merge channels from remaining datasets
-        for data in data_list[1:]:
-            channel_names.update(data.channel.values)  # add new channels
-
-        # Sort the merged list
-        ref_channel_names = sorted(channel_names, key=channel_sort_key)
-
-        ref = align_labels(ref_raw, ref_channel_names, x_dim).reindex(channel=ref_channel_names)
-
-        fig = plt.figure(figsize=(max(12, len(ref_channel_names) * 0.25), 6))
-
-        for data, run_id, lbl in zip(data_list, run_ids, label_list, strict=False):
-            if run_id == baseline_name:
-                continue  # skip baseline
-
-            num_raw = self._preprocess_data(data, x_dim, verbose=False)
-            num = align_labels(num_raw, ref_channel_names, x_dim).reindex(channel=ref_channel_names)
-
-            ratio = num.sel(channel=ref_channel_names) / ref.sel(channel=ref_channel_names)
-
-            plt.plot(
-                ref_channel_names,
-                ratio.values,
-                label=lbl,
-                marker="o",
-                linestyle="-",
-            )
-
-        parts = [descr, tag]
-        name = "_".join(filter(None, parts))
-        plt.xticks(rotation=90, ha="right")
-        plt.grid(True, linestyle="--", color="gray", alpha=0.2)
-        title = (
-            f"{descr.replace('_', ' ')} {tag.split('_')[0]} -"
-            f" {tag.split('_')[-1]} (baseline: {baseline_name})"
-        )
-        self._plot_base(fig, name, x_dim, y_dim, print_summary, line=1.0, vlines=True, title=title)
-
-    def heat_map(
-        self,
-        data: xr.DataArray | list,
-        labels: str | list,
-        metric: str,
-        x_dim,
-        tag: str = "",
-    ) -> None:
-        """
-        Plot a heat map comparing multiple datasets.
-        Parameters
-        ----------
-        data:
-            DataArray or list of DataArrays to be plotted
-        labels:
-            Label or list of labels for each dataset
-        metric:
-            Metric for which we are plotting
-        x_dim:
-            Dimension to be used for the x-axis. The code will average over all other dimensions.
-        tag:
-            Tag to be added to the plot title and filename
-        Returns
-        -------
-            None
-        """
-
-        data_list, label_list = self._check_lengths(data, labels)
-
-        n_runs = len(data_list)
-
-        x_ticks_names = set()
-
-        for data in data_list:
-            da = data.isel({x_dim: 0})
-            x_ticks_names.update(map(str, da.channel.values))
-
-        ref_ticks_names = sorted(x_ticks_names, key=channel_sort_key)
-
-        fig, axes = plt.subplots(
-            1, n_runs, figsize=(8 * n_runs, max(12, len(ref_ticks_names) * 0.25)), squeeze=False
-        )
-
-        global_min = float("inf")
-        global_max = float("-inf")
-
-        for ax, data, label in zip(axes[0], data_list, labels, strict=False):
-            time_steps = sorted(data[x_dim].values)
-
-            # Use the first time step as reference
-            ref = data.reindex(channel=ref_ticks_names).sel({x_dim: time_steps[0]})
-            ref = self._preprocess_data(ref, "channel", verbose=False)
-
-            if ref.isnull().all():
-                _logger.warning(
-                    f"Heatmap:: Reference data for metric {metric} and label {label} contains "
-                    "only NaNs. Skipping heatmap."
-                )
-                continue
-
-            # Compute ratio for all time steps
-            num = self._preprocess_data(data, [x_dim, "channel"], verbose=False)
-            num = num.reindex(channel=ref_ticks_names).sel({x_dim: time_steps})
-
-            heatmap_data = num / ref
-
-            cmap = plt.get_cmap("magma_r") if lower_is_better(metric) else plt.get_cmap("magma")
-            global_min = min(global_min, float(heatmap_data.min()))
-            global_max = max(global_max, float(heatmap_data.max()))
-
-            last_hm = sns.heatmap(
-                heatmap_data.values.T,
-                ax=ax,
-                cmap=cmap,
-                vmin=global_min,
-                vmax=global_max,
-                xticklabels=time_steps,
-                yticklabels=ref_ticks_names,
-                annot=False,
-                fmt=".2f",
-                cbar=False,
-            )
-            ax.set_title(f"Heatmap {metric} – {label}")
-            ax.set_xlabel(f"{x_dim.replace('_', ' ').title()} (h)")
-            ax.set_ylabel("Variable")
-            plt.setp(ax.get_xticklabels(), rotation=45, ha="right")
-
-        cbar = fig.colorbar(
-            last_hm.collections[0], ax=axes.ravel().tolist(), shrink=0.6, location="right", pad=0.02
-        )
-        cbar.set_label(rf"{metric} - $t_{{\mathrm{{step}}}}[0] / t_{{\mathrm{{step}}}}[x]$")
-        parts = ["heat_map", metric, tag]
-        name = "_".join(filter(None, parts))
-        plt.savefig(f"{self.out_plot_dir.joinpath(name)}.{self.image_format}")
-
-
-class QuantilePlots:
-    def __init__(self, plotter_cfg: dict, output_basedir: str | Path):
-        """
-        Initialize the QuantilePlots class.
-
-        Parameters
-        ----------
-        plotter_cfg:
-            Configuration dictionary containing basic information for plotting.
-            Expected keys are:
-                - image_format: Format of the saved images (e.g., 'png', 'pdf', etc.)
-                - dpi_val: DPI value for the saved images
-                - fig_size: Size of the figure (width, height) in inches
-        output_basedir:
-            Base directory under which the plots will be saved.
-            Expected scheme `<results_base_dir>/<run_id>`.
-        """
-        self.image_format = plotter_cfg.get("image_format")
-        self.dpi_val = plotter_cfg.get("dpi_val")
-        self.fig_size = plotter_cfg.get("fig_size")
-        self.out_plot_dir = Path(output_basedir) / "quantile_plots"
-
-        if not os.path.exists(self.out_plot_dir):
-            _logger.info(f"Creating dir {self.out_plot_dir}")
-            os.makedirs(self.out_plot_dir, exist_ok=True)
-
-    def _check_lengths(self, data: xr.DataArray | list, labels: str | list) -> tuple[list, list]:
-        """
-        Check if the lengths of data and labels match.
-
-        Parameters
-        ----------
-        data:
-            DataArray or list of DataArrays to be plotted
-        labels:
-            Label or list of labels for each dataset
-
-        Returns
-        -------
-            data_list, label_list - lists of data and labels
-        """
-        assert isinstance(data, xr.DataArray | list), (
-            "QuantilePlots::_check_lengths - Data should be of type xr.DataArray or list"
-        )
-        assert isinstance(labels, str | list), (
-            "QuantilePlots::_check_lengths - Labels should be of type str or list"
-        )
-
-        data_list = [data] if isinstance(data, xr.DataArray) else data
-        label_list = [labels] if isinstance(labels, str) else labels
-
-        assert len(data_list) == len(label_list), (
-            "QuantilePlots::_check_lengths - Data and Labels do not match"
-        )
-
-        return data_list, label_list
-
-    def qq_plot(
-        self,
-        qq_data: list[xr.Dataset],
-        labels: str | list,
-        tag: str = "",
-        metric: str = "qq_analysis",
-        extreme_percentiles: tuple[float, float] | None = None,
-    ) -> None:
-        """
-        Create quantile-quantile (Q-Q) plots for extreme value analysis.
-
-        This method generates comprehensive Q-Q plots comparing forecast quantiles
-        against ground truth quantiles, with emphasis on extreme values.
-
-        Parameters
-        ----------
-        qq_data:
-            Dataset or list of Datasets containing Q-Q analysis results.
-            Each dataset should contain:
-            - 'quantile_levels': Theoretical quantile levels (0 to 1)
-            - 'p_quantiles': Quantile values from prediction data
-            - 'gt_quantiles': Quantile values from ground truth data
-            - 'qq_deviation': Absolute difference between quantiles
-            - 'extreme_low_mse': MSE for lower extreme quantiles
-            - 'extreme_high_mse': MSE for upper extreme quantiles
-        labels:
-            Label or list of labels for each dataset
-        tag:
-            Tag to be added to the plot title and filename
-        metric:
-            Name of the metric (default: 'qq_analysis')
-        extreme_percentiles:
-            Lower and upper percentile thresholds for extreme regions.
-
-        Returns
-        -------
-            None
-        """
-        data_list, label_list = self._check_lengths(qq_data, labels)
-
-        # Use extreme_percentiles from data if not explicitly provided
-        if extreme_percentiles is None:
-            extreme_percentiles = tuple(data_list[0].attrs.get("extreme_percentiles", (5.0, 95.0)))
-
-        # Create figure with subplots
-        fig = plt.figure(figsize=(16, 6), dpi=self.dpi_val)
-        gs = fig.add_gridspec(1, 2, width_ratios=[2, 1], wspace=0.3)
-
-        ax_qq = fig.add_subplot(gs[0])  # Main Q-Q plot
-        ax_dev = fig.add_subplot(gs[1])  # Deviation plot
-
-        colors = plt.cm.tab10(np.linspace(0, 1, len(data_list)))
-
-        for _i, (ds, label, color) in enumerate(zip(data_list, label_list, colors, strict=False)):
-            # Extract quantile data
-            quantile_levels = ds["quantile_levels"].values
-            p_quantiles = ds["p_quantiles"].values
-            gt_quantiles = ds["gt_quantiles"].values
-            qq_deviation = ds["qq_deviation"].values
-
-            # Main Q-Q plot
-            ax_qq.scatter(
-                gt_quantiles,
-                p_quantiles,
-                alpha=0.6,
-                s=20,
-                c=[color],
-                label=label,
-                edgecolors="none",
-            )
-
-            # Deviation plot
-            ax_dev.plot(
-                quantile_levels,
-                qq_deviation,
-                label=label,
-                color=color,
-                linewidth=2,
-                alpha=0.8,
-            )
-
-        # Format main Q-Q plot
-        ax_qq.set_xlabel("Ground Truth Quantiles", fontsize=12)
-        ax_qq.set_ylabel("Prediction Quantiles", fontsize=12)
-        ax_qq.set_title("Quantile-Quantile Plot for Extreme Value Analysis", fontsize=14)
-
-        # Add perfect agreement line (y=x)
-        min_val = min([ds["gt_quantiles"].min().values for ds in data_list])
-        max_val = max([ds["gt_quantiles"].max().values for ds in data_list])
-        ax_qq.plot(
-            [min_val, max_val],
-            [min_val, max_val],
-            "k--",
-            linewidth=2,
-            label="Perfect Agreement",
-            alpha=0.7,
-        )
-
-        # Add shaded regions for extremes
-        if len(data_list) > 0:
-            ds_ref = data_list[0]
-            quantile_levels = ds_ref["quantile_levels"].values
-
-            # Find extreme regions using configurable thresholds
-            lower_extreme_idx = quantile_levels < (extreme_percentiles[0] / 100)
-            upper_extreme_idx = quantile_levels > (extreme_percentiles[1] / 100)
-
-            if np.any(lower_extreme_idx):
-                lower_q = ds_ref["gt_quantiles"].values[lower_extreme_idx]
-                ax_qq.axvspan(
-                    min_val,
-                    lower_q.max() if len(lower_q) > 0 else min_val,
-                    alpha=0.1,
-                    color="blue",
-                    label="Lower Extreme Zone",
-                )
-
-            if np.any(upper_extreme_idx):
-                upper_q = ds_ref["gt_quantiles"].values[upper_extreme_idx]
-                ax_qq.axvspan(
-                    upper_q.min() if len(upper_q) > 0 else max_val,
-                    max_val,
-                    alpha=0.1,
-                    color="red",
-                    label="Upper Extreme Zone",
-                )
-
-        ax_qq.legend(frameon=False, loc="upper left", fontsize=10)
-        ax_qq.grid(True, linestyle="--", alpha=0.3)
-
-        # Format deviation plot
-        ax_dev.set_xlabel("Quantile Level", fontsize=12)
-        ax_dev.set_ylabel("Absolute Deviation", fontsize=12)
-        ax_dev.set_title("Quantile Deviation", fontsize=14)
-        ax_dev.legend(frameon=False, fontsize=10)
-        ax_dev.grid(True, linestyle="--", alpha=0.3)
-
-        # Highlight extreme regions in deviation plot
-        lower_threshold = extreme_percentiles[0] / 100
-        upper_threshold = extreme_percentiles[1] / 100
-        ax_dev.axvspan(0.0, lower_threshold, alpha=0.1, color="blue")
-        ax_dev.axvspan(upper_threshold, 1.0, alpha=0.1, color="red")
-
-        plt.tight_layout()
-
-        # Save the plot
-        parts = ["qq_analysis", tag]
-        name = "_".join(filter(None, parts))
-        save_path = self.out_plot_dir.joinpath(f"{name}.{self.image_format}")
-        plt.savefig(save_path, bbox_inches="tight")
-        plt.close()
-
-        _logger.info(f"Q-Q plot saved to {save_path}")
-
-
-class ScoreCards:
-    """
-    Initialize the ScoreCards class.
-
-    Parameters
-    ----------
-    plotter_cfg:
-        Configuration dictionary containing basic information for plotting.
-        Expected keys are:
-            - image_format: Format of the saved images (e.g., 'png', 'pdf', etc.)
-            - improvement: Size of the figure (width, height) in inches
-    output_basedir:
-        Base directory under which the score cards will be saved.
-    """
-
-    def __init__(self, plotter_cfg: dict, output_basedir: str | Path) -> None:
-        self.image_format = plotter_cfg.get("image_format")
-        self.dpi_val = plotter_cfg.get("dpi_val")
-        self.improvement = plotter_cfg.get("improvement_scale", 0.2)
-        self.out_plot_dir = Path(output_basedir) / "score_cards"
-        self.baseline = plotter_cfg.get("baseline")
-        if not os.path.exists(self.out_plot_dir):
-            _logger.info(f"Creating dir {self.out_plot_dir}")
-            os.makedirs(self.out_plot_dir, exist_ok=True)
-
-    def plot(
-        self,
-        data: list[xr.DataArray],
-        runs: list[str],
-        metric: str,
-        channels: list[str],
-        tag: str,
-    ) -> None:
-        """
-        Plot score cards comparing performance between run_ids against a baseline over channels
-        of interest.
-
-        Parameters
-        ----------
-        data:
-            List of (xarray) DataArrays with the scores (stream, region and metric specific)
-        runs:
-            List containing runs (in str format) to be compared (provided in the config)
-        metric:
-            Metric for which we are plotting
-        channels:
-            List containing channels (in str format) of interest (provided in the config)
-        tag:
-            Tag to be added to the plot title and filename
-        """
-        n_runs = len(runs)
-
-        if self.baseline and self.baseline in runs:
-            baseline_idx = runs.index(self.baseline)
-            runs = [runs[baseline_idx]] + runs[:baseline_idx] + runs[baseline_idx + 1 :]
-            data = [data[baseline_idx]] + data[:baseline_idx] + data[baseline_idx + 1 :]
-
-        common_channels, n_common_channels = self.extract_common_channels(data, channels, n_runs)
-
-        fig, ax = plt.subplots(figsize=(2 * n_runs, 1.2 * n_common_channels))
-
-        baseline = data[0]
-        skill_models = []
-        for run_index in range(1, n_runs):
-            skill_model = 0.0
-            data0_channels = [str(x) for x in np.atleast_1d(data[0].channel.values)]
-            data_idx_channels = [str(x) for x in np.atleast_1d(data[run_index].channel.values)]
-            for var_index, var in enumerate(common_channels):
-                if var not in data0_channels or var not in data_idx_channels:
-                    continue
-                diff, avg_diff, avg_skill = self.compare_models(
-                    data, baseline, run_index, var, metric
-                )
-                skill_model += avg_skill.values
-
-                # Get symbols based on difference and performance as well as coordinates
-                # for the position of the triangles.
-
-                x, y, alt, color, triangle, size = self.get_plot_symbols(
-                    run_index, var_index, avg_skill, avg_diff, metric
-                )
-
-                ax.scatter(x, y, marker=triangle, color=color, s=size.values, zorder=3)
-
-                # Perform Wilcoxon test
-                if len(diff["forecast_step"].values) > 1:
-                    stat, p = wilcoxon(diff, alternative=alt)
-
-                    # Draw rectangle border for significance
-                    if p < 0.05:
-                        lw = 2 if p < 0.01 else 1
-                        rect_color = color
-                        rect = plt.Rectangle(
-                            (x - 0.25, y - 0.25),
-                            0.5,
-                            0.5,
-                            fill=False,
-                            edgecolor=rect_color,
-                            linewidth=lw,
-                            zorder=2,
-                        )
-                        ax.add_patch(rect)
-
-            skill_models.append(skill_model / n_common_channels)
-
-        # Set axis labels
-        ylabels = [
-            f"{var}\n({baseline.coords['metric'].item().upper()}={baseline.sel(channel=var).mean().values.squeeze():.3f})"
-            for var in common_channels
-        ]
-        xlabels = [
-            f"{model_name}\nSkill: {skill_models[i]:.3f}" for i, model_name in enumerate(runs[1::])
-        ]
-        ax.set_xticks(np.arange(1, n_runs))
-        ax.set_xticklabels(xlabels, fontsize=10)
-        ax.set_yticks(np.arange(n_common_channels) + 0.5)
-        ax.set_yticklabels(ylabels, fontsize=10)
-        for label in ax.get_yticklabels():
-            label.set_horizontalalignment("center")
-            label.set_x(-0.17)
-        ax.set_ylabel("Variable", fontsize=14)
-        ax.set_title(
-            f"Model Scorecard vs. Baseline '{runs[0]}'",
-            fontsize=16,
-            pad=20,
-        )
-        for x in np.arange(0.5, n_runs - 1, 1):
-            ax.axvline(x, color="gray", linestyle="--", linewidth=0.5, zorder=0, alpha=0.5)
-        ax.set_xlim(0.5, n_runs - 0.5)
-        ax.set_ylim(0, n_common_channels)
-
-        legend = [
-            Line2D(
-                [0],
-                [0],
-                marker="^",
-                color="white",
-                label=f"{self.improvement * 100:.0f}% improvement",
-                markerfacecolor="blue",
-                markersize=np.sqrt(200),
-            )
-        ]
-        plt.legend(handles=legend, loc="upper left", bbox_to_anchor=(1.02, 1.0))
-
-        _logger.info(f"Saving scorecards to: {self.out_plot_dir}")
-
-        parts = ["score_card", tag] + runs
-        name = "_".join(filter(None, parts))
-        plt.savefig(
-            f"{self.out_plot_dir.joinpath(name)}.{self.image_format}",
-            bbox_inches="tight",
-            dpi=self.dpi_val,
-        )
-        plt.close(fig)
-
-    def extract_common_channels(self, data, channels, n_runs):
-        common_channels = []
-        for run_index in range(1, n_runs):
-            data0_channels = [str(x) for x in np.atleast_1d(data[0].channel.values)]
-            data_idx_channels = [str(x) for x in np.atleast_1d(data[run_index].channel.values)]
-            for var in channels:
-                if var not in data0_channels or var not in data_idx_channels:
-                    continue
-                common_channels.append(var)
-        common_channels = list(set(common_channels))
-        n_vars = len(common_channels)
-        return common_channels, n_vars
-
-    def compare_models(
-        self,
-        data: list[xr.DataArray],
-        baseline: xr.DataArray,
-        run_index: int,
-        var: str,
-        metric: str,
-        x_dim="forecast_step",
-    ) -> tuple[xr.DataArray, xr.DataArray, xr.DataArray]:
-        """
-        Compare a model with a baseline model and calculate skill scores.
-
-        Parameters
-        ----------
-        data: list[xr.DataArray]
-            List of all scores in xarray format for each model.
-
-        baseline: xarray DataArray
-            The baseline scores in xarrays format.
-
-        run_index: int
-            The order index over the run_ids.
-
-        var: str
-            The specified channel over which we compare.
-
-        xdim: str
-            The dimension for which an average will not be calculated.
-
-        Returns
-        ----------
-        diff: xr.DataArray
-            Difference in scores between baseline and model.
-
-        diff.mean(dim="forecast_step"): xr.DataArray
-            Average difference in scores over all forecast steps between baseline and model .
-
-        skill.mean(dim="forecast_step"): xr.DataArray
-            Average skill scores over all forecast steps between baseline and model .
-
-        """
-        baseline_var = baseline.sel({"channel": var})
-        data_var = data[run_index].sel({"channel": var})
-
-        baseline_score, model_score = calculate_average_over_dim(x_dim, baseline_var, data_var)
-        diff = baseline_score - model_score
-
-        skill = self.get_skill_score(model_score, baseline_score, metric)
-        return diff, diff.mean(dim=x_dim), skill.mean(dim=x_dim)
-
-    def get_skill_score(
-        self, score_model: xr.DataArray, score_ref: xr.DataArray, metric: str
-    ) -> xr.DataArray:
-        """
-        Calculate skill score comparing a model against a baseline.
-
-        Skill score is defined as: (model_score - baseline_score) / (perfect_score - baseline_score)
-
-        Parameters
-        ----------
-        score_model : xr.DataArray
-            The scores of the model being evaluated
-        score_ref : xr.DataArray
-            The scores of the reference/baseline model
-        metric : str
-            The metric name for which to calculate skill score
-
-        Returns
-        -------
-        xr.DataArray
-            Skill scores comparing model to baseline
-        """
-        perf_score = self.get_perf_score(metric)
-        skill_score = (score_model - score_ref) / (perf_score - score_ref)
-        return skill_score
-
-    def get_perf_score(self, metric: str) -> float:
-        """
-        Get the perfect score for a given metric.
-
-        Perfect scores represent ideal performance:
-        - Error metrics: 0 (lower is better)
-        - Skill/score metrics: 1 (higher is better)
-        - PSNR: 100 (higher is better)
-
-        Parameters
-        ----------
-        metric : str
-            Metric name
-
-        Returns
-        -------
-        float
-            Perfect score for the specified metric
-        """
-        # Metrics where lower values indicate better performance (error metrics)
-        if lower_is_better(metric):
-            return 0.0
-
-        # Metrics where higher values indicate better performance (with specific perfect score)
-        elif metric in ["psnr"]:
-            return 100.0
-
-        # Metrics where higher values indicate better performance (default perfect score)
-        else:
-            return 1.0
-
-    def get_plot_symbols(
-        self,
-        run_index: int,
-        var_index: int,
-        avg_skill: xr.DataArray,
-        avg_diff: xr.DataArray,
-        metric: str,
-    ) -> tuple[int, float, str, str, str, xr.DataArray]:
-        """
-        Determine plot symbol properties based on performance difference.
-
-        Parameters
-        ----------
-        run_index : int
-            Index of the model.
-        var_index : int
-            Index of the variable/channel.
-        avg_skill : xr.DataArray
-            Average skill score of the model.
-        avg_diff : xr.DataArray
-            Average difference between baseline and model.
-        metric : str
-            Metric used for interpretation.
-
-        Returns
-        -------
-        Tuple[int, float, str, str, str, xr.DataArray]
-            x, y coordinates, alternative hypothesis, color, triangle symbol, size.
-        """
-        # Conservative choice
-        alt = "two-sided"
-        modus = "different"
-        color = "gray"
-
-        # Determine if diff_mean indicates improvement
-        is_improvement = (avg_diff > 0 and lower_is_better(metric)) or (
-            avg_diff < 0 and not lower_is_better(metric)
-        )
-
-        if is_improvement:
-            alt = "greater"
-            modus = "better"
-            color = "blue"
-        elif not is_improvement and avg_diff != 0:
-            alt = "less"
-            modus = "worse"
-            color = "red"
-        else:
-            alt = "two-sided"
-            modus = "different"
-
-        triangle = "^" if modus == "better" else "v"
-
-        # Triangle coordinates
-        x = run_index
-        # First row is model 1 vs model 0
-        y = var_index + 0.5
-
-        size = 200 * (1 - (1 / (1 + abs(avg_skill) / self.improvement)))  # Add base size to all
-
-        return x, y, alt, color, triangle, size
-
-
-class BarPlots:
-    """
-    Initialize the BarPlots class.
-
-    Parameters
-    ----------
-    plotter_cfg:
-        Configuration dictionary containing basic information for plotting.
-        Expected keys are:
-            - image_format: Format of the saved images (e.g., 'png', 'pdf', etc.)
-            - improvement: Size of the figure (width, height) in inches
-    output_basedir:
-        Base directory under which the score cards will be saved.
-    """
-
-    def __init__(self, plotter_cfg: dict, output_basedir: str | Path) -> None:
-        self.image_format = plotter_cfg.get("image_format")
-        self.dpi_val = plotter_cfg.get("dpi_val")
-        self.cmap = plotter_cfg.get("cmap", "bwr")
-        self.out_plot_dir = Path(output_basedir) / "bar_plots"
-        self.baseline = plotter_cfg.get("baseline")
-        if not os.path.exists(self.out_plot_dir):
-            _logger.info(f"Creating dir {self.out_plot_dir}")
-            os.makedirs(self.out_plot_dir, exist_ok=True)
-
-    def plot(
-        self,
-        data: list[xr.DataArray],
-        runs: list[str],
-        metric: str,
-        channels: list[str],
-        tag: str,
-    ) -> None:
-        """
-        Plot (ratio) bar plots comparing performance between different run_ids over channels of
-        interest.
-
-        Parameters
-        ----------
-        data:
-            List of (xarray) DataArrays with the scores (stream, region and metric specific)
-        runs:
-            List containing runs (in str format) to be compared (provided in the config)
-        metric:
-            Metric name
-        channels:
-            List containing channels (in str format) of interest (provided in the config)
-        tag:
-            Tag to be added to the plot title and filename
-        """
-
-        fig, ax = plt.subplots(
-            1,
-            len(runs) - 1 if len(runs) > 1 else 1,
-            figsize=(5 * len(runs), 2 * len(channels)),
-            dpi=self.dpi_val,
-            squeeze=False,
-        )
-        ax = ax.flatten()
-        if self.baseline and self.baseline in runs:
-            baseline_idx = runs.index(self.baseline)
-            runs = [runs[baseline_idx]] + runs[:baseline_idx] + runs[baseline_idx + 1 :]
-            data = [data[baseline_idx]] + data[:baseline_idx] + data[baseline_idx + 1 :]
-        elif len(runs) < 2:
-            _logger.warning(
-                "BarPlots:: Less than two runs provided. Generating bar plot against ones."
-            )
-            ones_array = xr.full_like(data[0], 1.0)
-            runs = [""] + runs
-            data = [ones_array] + data
-
-        for run_index in range(1, len(runs)):
-            score, channels_per_comparison = self.calc_ratio_per_run_id(data, channels, run_index)
-            if len(score) > 0:
-                ax[run_index - 1].barh(
-                    np.arange(len(score)),
-                    score,
-                    color=self.colors(score, metric),
-                    align="center",
-                    edgecolor="black",
-                    linewidth=0.5,
-                )
-                ax[run_index - 1].set_yticks(np.arange(len(score)), labels=channels_per_comparison)
-                ax[run_index - 1].invert_yaxis()
-
-                xlabel = (
-                    f"Relative {data[0].coords['metric'].item().upper()}: "
-                    f"Target Model ({runs[run_index]}) / Reference Model ({runs[0]})"
-                )
-
-                if len(runs) == 2 and runs[0] == "":
-                    xlabel = xlabel.replace("Relative ", "")
-                    xlabel = xlabel.replace(
-                        f"Target Model ({runs[run_index]}) / Reference Model ({runs[0]})",
-                        f"Model ({runs[run_index]})",
-                    )
-
-                ax[run_index - 1].set_xlabel(xlabel)
-            else:
-                ax[run_index - 1].set_visible(False)  # or annotate as missing
-                # Or show a message:
-                ax[run_index - 1].text(
-                    0.5,
-                    0.5,
-                    "No Data",
-                    ha="center",
-                    va="center",
-                    transform=ax[run_index - 1].transAxes,
-                )
-
-        _logger.info(f"Saving bar plots to: {self.out_plot_dir}")
-        parts = ["bar_plot", tag] + runs
-        name = "_".join(filter(None, parts))
-        plt.savefig(
-            f"{self.out_plot_dir.joinpath(name)}.{self.image_format}",
-            bbox_inches="tight",
-            dpi=self.dpi_val,
-        )
-        plt.close(fig)
-
-    def calc_ratio_per_run_id(
-        self,
-        data: list[xr.DataArray],
-        channels: list[str],
-        run_index: int,
-        x_dim="channel",
-    ) -> tuple[np.array, str]:
-        """
-        This function calculates the ratio per comparison model for each channel.
-
-        Parameters
-        ----------
-        data: list[xr.DataArray]
-            List of all scores for each model in xarrays format.
-        channels: list[str]
-            All the available channels.
-        run_index: int
-            The order index over the run_ids.
-        xdim: str
-            The dimension for which an average will not be calculated.
-
-        Returns
-        ----------
-        ratio_score: np.array
-            The (ratio) skill over each channel for a specific model
-        channels_per_comparison: str
-            The common channels over which the baseline and the other model will be compared.
-
-        """
-        ratio_score = []
-        channels_per_comparison = []
-
-        for _, var in enumerate(channels):
-            if var not in data[0].channel.values or var not in data[run_index].channel.values:
-                continue
-            baseline_var = data[0].sel({"channel": var})
-            data_var = data[run_index].sel({"channel": var})
-            channels_per_comparison.append(var)
-
-            baseline_score, model_score = calculate_average_over_dim(x_dim, baseline_var, data_var)
-
-            ratio_score.append(model_score / baseline_score)
-
-        if np.allclose(baseline_score, 1.0, atol=1e-6):
-            ratio_score = np.array(ratio_score)
-        else:
-            ratio_score = np.array(ratio_score) - 1
-
-        return ratio_score, channels_per_comparison
-
-    def colors(self, ratio_score: np.array, metric: str) -> list[tuple]:
-        """
-        This function calculates colormaps based on the skill scores. From negative value blue
-        color variations should be given otherwise red color variations should be given.
-
-        Parameters
-        ----------
-        ratio_score: np.array
-            The (ratio) skill for a specific model
-        metric: str
-            The metric of interest
-        Returns
-        ----------
-        colors: list[tuple]
-            The color magnitude (blue to red) of the bars in the plots
-        """
-        max_val = np.abs(ratio_score).max()
-        if lower_is_better(metric):
-            cmap = plt.get_cmap("bwr")
-        else:
-            cmap = plt.get_cmap("bwr_r")
-        colors = [cmap(0.5 + v / (2 * max_val)) for v in ratio_score]
-        return colors
-
-
-def calculate_average_over_dim(
-    x_dim: str, baseline_var: xr.DataArray, data_var: xr.DataArray
-) -> tuple[xr.DataArray, xr.DataArray]:
-    """
-    Calculate average over xarray dimensions that are larger than 1. Those might be the
-    forecast-steps or the samples.
-
-    Parameters
-    ----------
-    xdim: str
-        The dimension for which an average will not be calculated.
-    baseline_var: xr.DataArray
-        xarray DataArray with the scores of the baseline model for a specific channel/variable
-    data_var: xr.DataArray
-        xarray DataArray with the scores of the comparison model for a specific channel/variable
-
-    Returns
-    -------
-    baseline_score: xarray DataArray
-        The baseline average scores over the dimensions not specified by xdim
-    model_score: xarray DataArray
-        The model average scores over the dimensions not specified by xdim
-    """
-    non_zero_dims = [
-        dim for dim in baseline_var.dims if dim != x_dim and baseline_var[dim].shape[0] > 1
-    ]
-
-    if non_zero_dims:
-        _logger.info(f"Found multiple entries for dimensions: {non_zero_dims}. Averaging...")
-
-    baseline_score = baseline_var.mean(
-        dim=[dim for dim in baseline_var.dims if dim != x_dim], skipna=True
-    )
-    model_score = data_var.mean(dim=[dim for dim in data_var.dims if dim != x_dim], skipna=True)
-
-    return baseline_score, model_score
-
-
-def lower_is_better(metric: str) -> bool:
-    # Determine whether lower or higher is better
-    return metric in {"l1", "l2", "mae", "mse", "rmse", "vrmse", "bias", "crps", "spread"}
-
-
-def compute_offsets(n, spacing=0.11):
-    idx = np.arange(n)
-    return (idx - (n - 1) / 2.0) * spacing
-
-
-def align_labels(da: xr.DataArray, labels: list[str], x_dim: str) -> xr.DataArray:
-    """
-    Reindex a DataArray to include all labels in the canonical order.
-    Missing variables are filled with NaN.
-    """
-    # Convert labels → index format expected by xarray
-    labels = np.array(labels, dtype=object)
-
-    # Reindex, inserting NaN for missing labels
-    return da.reindex({x_dim: labels})
-
-
-def format_datetime(dt):
-    return dt.astype("datetime64[m]").astype(datetime.datetime).strftime("%Y-%m-%d T%H:%M:%S")
-
-
-def channel_sort_key(name: str) -> tuple[int, str, int]:
-    """
-    Sorting key for channel names like 't_850', 'z_500', etc.
-    Splits the name into a prefix and a number suffix for sorting.
-    Parameters
-    ----------
-    name : str
-        Channel name to be sorted.
-    Returns
-    -------
-    tuple[int, str, int]
-        Sorting key: (0, prefix, number) if pattern matches, else (1,
-    """
-    m = re.match(r"(.+?)_(\d+)$", name)
-    if m:
-        prefix, number = m.groups()
-        return (0, prefix, int(number))
-    else:
-        return (1, name, float("inf"))
